requirements.txt
openai>=1.40.0
watchdog>=4.0.0
typer>=0.12.3
pydantic>=2.8.2
python-dotenv>=1.0.1
PyYAML>=6.0.2
Pillow>=10.4.0
rich>=13.9.1




